+++ internal/handler/management/constants.go
@@ -1,4 +1,4 @@
-// Copyright 2022 Dimitrij Drus <dadrus@gmx.de>
+// Copyright 2023 Dimitrij Drus <dadrus@gmx.de>
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,28 +14,9 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package management2
+package management
 
-import (
-	"net/http"
-
-	"github.com/goccy/go-json"
+const (
+	EndpointHealth = "/.well-known/health"
+	EndpointJWKS   = "/.well-known/jwks"
 )
-
-func health() http.Handler {
-	type status struct {
-		Status string `json:"status"`
-	}
-
-	return http.HandlerFunc(func(rw http.ResponseWriter, _ *http.Request) {
-		res, err := json.Marshal(status{Status: "ok"})
-		if err != nil {
-			rw.WriteHeader(http.StatusInternalServerError)
-
-			return
-		}
-
-		rw.Header().Set("Content-Type", "application/json")
-		rw.Write(res)
-	})
-}
