+++ internal/handler/management/handler.go
@@ -1,4 +1,4 @@
-package management2
+package management
 
 import (
 	"net/http"
